extract BenchmarkAnswerExplanationGraph
